🧹 extract range from llx
